Move annotate by interval to annotate_gnomad.py
